requirements.txt
torch>=2.0.0
torch-geometric
pandas
numpy
scikit-learn
pyyaml
tqdm
tensorboard
